Implementing faculty aggregate page
