+++ uni/lib/utils/drawer_items.dart
@@ -8,7 +8,7 @@ enum DrawerItem {
   navRestaurants('restaurantes'),
   navCalendar('calendario'),
   navLibrary('biblioteca', faculties: {'feup'}),
-  navUsefulInfo('uteis', faculties: {'feup'});
+  navFaculty('faculdade');
 
   const DrawerItem(this.title, {this.faculties});
