+++ uni/lib/view/faculty/widgets/academic_services_card.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:uni/generated/l10n.dart';
 import 'package:uni/utils/drawer_items.dart';
 import 'package:uni/view/common_widgets/generic_expansion_card.dart';
-import 'package:uni/view/useful_info/widgets/text_components.dart';
+import 'package:uni/view/faculty/widgets/text_components.dart';
 
-class InfoDeskCard extends GenericExpansionCard {
-  const InfoDeskCard({super.key});
+class AcademicServicesCard extends GenericExpansionCard {
+  const AcademicServicesCard({super.key});
 
   @override
   Widget buildCardContent(BuildContext context) {
@@ -16,21 +16,21 @@ class InfoDeskCard extends GenericExpansionCard {
           context,
           initial: true,
         ),
-        h2(S.of(context).tele_personal_assistance, context),
-        infoText('9:30h - 13:00h | 14:00h - 17:30h', context),
+        h2(S.of(context).personal_assistance, context),
+        infoText('11:00h - 16:00h', context),
+        h2(S.of(context).tele_assistance, context),
+        infoText('9:30h - 12:00h | 14:00h - 16:00h', context),
         h1(S.of(context).telephone, context),
-        infoText('+351 225 081 400', context, link: 'tel:225 081 400'),
-        h1('Email', context),
         infoText(
-          'infodesk@fe.up.pt',
+          '+351 225 081 977',
           context,
+          link: 'tel:225 081 977',
           last: true,
-          link: 'mailto:infodesk@fe.up.pt',
         ),
       ],
     );
   }
 
   @override
-  String getTitle(BuildContext context) => 'Infodesk';
+  String getTitle(BuildContext context) => S.of(context).academic_services;
 }
