+++ uni/lib/view/faculty/widgets/copy_center_card.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:uni/generated/l10n.dart';
 import 'package:uni/utils/drawer_items.dart';
 import 'package:uni/view/common_widgets/generic_expansion_card.dart';
-import 'package:uni/view/useful_info/widgets/text_components.dart';
+import 'package:uni/view/faculty/widgets/text_components.dart';
 
 class CopyCenterCard extends GenericExpansionCard {
   const CopyCenterCard({super.key});
