+++ uni/lib/view/faculty/widgets/dona_bia_card.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:uni/generated/l10n.dart';
 import 'package:uni/utils/drawer_items.dart';
 import 'package:uni/view/common_widgets/generic_expansion_card.dart';
-import 'package:uni/view/useful_info/widgets/text_components.dart';
+import 'package:uni/view/faculty/widgets/text_components.dart';
 
-class AcademicServicesCard extends GenericExpansionCard {
-  const AcademicServicesCard({super.key});
+class DonaBiaCard extends GenericExpansionCard {
+  const DonaBiaCard({super.key});
 
   @override
   Widget buildCardContent(BuildContext context) {
@@ -16,21 +16,21 @@ class AcademicServicesCard extends GenericExpansionCard {
           context,
           initial: true,
         ),
-        h2(S.of(context).personal_assistance, context),
-        infoText('11:00h - 16:00h', context),
-        h2(S.of(context).tele_assistance, context),
-        infoText('9:30h - 12:00h | 14:00h - 16:00h', context),
+        h2(S.of(context).dona_bia_building, context),
+        infoText('8:30h - 12:00h | 13:30h - 19:00h', context),
         h1(S.of(context).telephone, context),
+        infoText('+351 225 081 416', context, link: 'tel:225 081 416'),
+        h1('Email', context),
         infoText(
-          '+351 225 081 977',
+          'papelaria.fe.up@gmail.com',
           context,
-          link: 'tel:225 081 977',
           last: true,
+          link: 'mailto:papelaria.fe.up@gmail.com',
         ),
       ],
     );
   }
 
   @override
-  String getTitle(BuildContext context) => S.of(context).academic_services;
+  String getTitle(BuildContext context) => S.of(context).dona_bia;
 }
