+++ uni/lib/view/faculty/widgets/infodesk_card.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:uni/generated/l10n.dart';
 import 'package:uni/utils/drawer_items.dart';
 import 'package:uni/view/common_widgets/generic_expansion_card.dart';
-import 'package:uni/view/useful_info/widgets/text_components.dart';
+import 'package:uni/view/faculty/widgets/text_components.dart';
 
-class MultimediaCenterCard extends GenericExpansionCard {
-  const MultimediaCenterCard({super.key});
+class InfoDeskCard extends GenericExpansionCard {
+  const InfoDeskCard({super.key});
 
   @override
   Widget buildCardContent(BuildContext context) {
@@ -16,21 +16,21 @@ class MultimediaCenterCard extends GenericExpansionCard {
           context,
           initial: true,
         ),
-        h2('${S.of(context).room} B123', context),
-        infoText('9:00h - 12:30h | 14:30h - 17:00h', context),
+        h2(S.of(context).tele_personal_assistance, context),
+        infoText('9:30h - 13:00h | 14:00h - 17:30h', context),
         h1(S.of(context).telephone, context),
-        infoText('+351 225 081 466', context, link: 'tel:225 081 466'),
+        infoText('+351 225 081 400', context, link: 'tel:225 081 400'),
         h1('Email', context),
         infoText(
-          'imprimir@fe.up.pt',
+          'infodesk@fe.up.pt',
           context,
           last: true,
-          link: 'mailto:imprimir@fe.up.pt',
+          link: 'mailto:infodesk@fe.up.pt',
         ),
       ],
     );
   }
 
   @override
-  String getTitle(BuildContext context) => S.of(context).multimedia_center;
+  String getTitle(BuildContext context) => 'Infodesk';
 }
