+++ uni/lib/view/faculty/widgets/sigarra_links_card.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:uni/generated/l10n.dart';
 import 'package:uni/view/common_widgets/generic_expansion_card.dart';
-import 'package:uni/view/useful_info/widgets/link_button.dart';
+import 'package:uni/view/faculty/widgets/link_button.dart';
 
 /// Manages the 'Current account' section inside the user's page (accessible
 /// through the top-right widget with the user picture)
