+++ uni/lib/view/faculty/widgets/text_components.dart
